refactor(mso_mdoc): split cred_processor.py into signing_key and payload modules
Extract standalone functions from cred_processor.py into focused modules:

- signing_key.py: check_certificate_not_expired + resolve_signing_key_for_credential
- payload.py: prepare_mdoc_payload + normalize_mdoc_result

cred_processor.py retains MsoMdocCredProcessor and re-exports the extracted
public names. Private methods _prepare_payload and _normalize_mdoc_result remain
as one-liner delegates to preserve the existing test API.

Update 6 test patch paths from mso_mdoc.cred_processor.MdocStorageManager to
mso_mdoc.signing_key.MdocStorageManager for tests of the standalone
resolve_signing_key_for_credential function; update debug-log capture logger
from mso_mdoc.cred_processor to mso_mdoc.payload.
